requirements.txt
lark-oapi>=1.4.0
openai>=1.0.0
python-dotenv>=1.0.0
